add android support
The implementation for android uses ptytty.c like sunos because Android
NDK does not have libutil. But in ptytty.c, 'STREAMS modules pushing' is
not supported in Android and seems to be not needed, therefore I disable
it for Android.
And in proc.c, Android does not support WIFCONTINUED, it uses WIFSTOPPED
and WSTOPSIG instead.

NOTICE:
Some environment variables are needed for make. This notice is also
written in make_android.mak with an example.