Allow `make` to autodetect the platform
If this gets more complicated automake/autoconf should be used to
detect the platform and configure the build.

This makefile simply checks the platform and then will invoke the
appropriate makefile for that platform. Things could probably be
condensed into a single makefile, at least for all the *nix
platforms.