Use clang to build vimproc in Mac
I changed to use clang to build vimproc in Mac because llvm-gcc
seems to be removed in Marvericks.

I ran tests and confirmed that all tests are passed except for
test-fopen, which is not passed even if llvm-gcc is used.
Additionally, I use some plugins which uses vimproc with vimproc built
with clang.  It seems to work well.